Move allocators code to separate files
